Refactored community card components
