+++ packages/commonwealth/client/scripts/views/components/CommunityCard/CommunityCard.scss
@@ -1,4 +1,4 @@
-@import '../shared';
+@import '../../../../styles/shared';
 
 $large-chain-icon-size: 96px;
 $small-chain-icon-size: 40px;
