+++ packages/commonwealth/client/scripts/views/components/CommunityCard/index.ts
@@ -1,0 +1,1 @@
+export { CommunityCard, NewCommunityCard } from './CommunityCard';
